Task 227 work inventory logic

See merge request StanislavKosyakov/unionandroid!395

+++ app/src/main/kotlin/com/itrocket/union/inventoryCreate/presentation/view/InventoryCreateScreen.kt
@@ -244,10 +244,10 @@ private fun Content(
 
 @Composable
 fun CountBar(
-    allAccountingObjects: Int,
-    findAccountingObjects: Int,
-    notFindAccountingObjects: Int,
-    newAccountingObjects: Int
+    allAccountingObjects: Long,
+    findAccountingObjects: Long,
+    notFindAccountingObjects: Long,
+    newAccountingObjects: Long
 ) {
     Row(
         modifier = Modifier
